+++ packages/block-library/src/search/index.php
@@ -188,25 +188,51 @@ function styles_for_block_core_search( $attributes ) {
 	$has_border_radius = ! empty( $attributes['style']['border']['radius'] );
 
 	if ( $has_border_radius ) {
-		// Shared style for button and input radius values.
+		$default_padding = '4px';
 		$border_radius   = $attributes['style']['border']['radius'];
-		$border_radius   = is_numeric( $border_radius ) ? $border_radius . 'px' : $border_radius;
-		$shared_styles[] = sprintf( 'border-radius: %s;', esc_attr( $border_radius ) );
-
-		// Apply wrapper border radius if button placed inside.
-		$button_inside = ! empty( $attributes['buttonPosition'] ) &&
+		$button_inside   = ! empty( $attributes['buttonPosition'] ) &&
 			'button-inside' === $attributes['buttonPosition'];
 
-		if ( $button_inside ) {
-			// We adjust the border radius value for the outer wrapper element
-			// to make it visually consistent with the radius applied to inner
-			// elements. calc() is used to support non-pixel CSS units.
-			$default_padding  = '4px';
-			$wrapper_styles[] = sprintf(
-				'border-radius: calc(%s + %s);',
-				esc_attr( $border_radius ),
-				esc_attr( $default_padding )
-			);
+		if ( is_array( $border_radius ) ) {
+			// Apply styles for individual corner border radii.
+			foreach ( $border_radius as $key => $value ) {
+				if ( null !== $value ) {
+					// Convert camelCase key to kebab-case.
+					$name = strtolower( preg_replace( '/(?<!^)[A-Z]/', '-$0', $key ) );
+
+					// Add shared styles for individual border radii for input & button.
+					$shared_styles[] = sprintf(
+						'border-%s-radius: %s;',
+						esc_attr( $name ),
+						esc_attr( $value )
+					);
+
+					// Add adjusted border radius styles for the wrapper element
+					// if button is positioned inside.
+					if ( $button_inside && intval( $value ) !== 0 ) {
+						$wrapper_styles[] = sprintf(
+							'border-%s-radius: calc(%s + %s);',
+							esc_attr( $name ),
+							esc_attr( $value ),
+							$default_padding
+						);
+					}
+				}
+			}
+		} else {
+			// Numeric check is for backwards compatibility purposes.
+			$border_radius   = is_numeric( $border_radius ) ? $border_radius . 'px' : $border_radius;
+			$shared_styles[] = sprintf( 'border-radius: %s;', esc_attr( $border_radius ) );
+
+			if ( $button_inside && intval( $border_radius ) !== 0 ) {
+				// Adjust wrapper border radii to maintain visual consistency
+				// with inner elements when button is positioned inside.
+				$wrapper_styles[] = sprintf(
+					'border-radius: calc(%s + %s);',
+					esc_attr( $border_radius ),
+					$default_padding
+				);
+			}
 		}
 	}