+++ packages/block-library/CHANGELOG.md
@@ -7,6 +7,10 @@
 -   Removes the `core/legacy-widget` block. This is now in `@wordpress/widgets`
 	via `registerLegacyWidgetBlock()`.
 
+## Bug Fixes
+
+-   Include missing attributes when upgrading embed block ([#33235](https://github.com/WordPress/gutenberg/pull/33235))
+
 ## 3.2.0 (2021-05-24)
 
 ### New Features